Matlab process to handle data
This relies on manual visual matching of data lines to ensure robot arm and imudata align. 
Once matching lines are entered, the process is somewhat automatic at generating Kalman and Madgwick data 
This code is used as a base to handle rolling-only data but filenames can be changed and it will process other experiments as well.